+++ packages/log/index.js
@@ -15,9 +15,9 @@ const defaults = {
     base: null,
     enabled: process.env.LOG_ENABLED !== 'false',
     level: process.env.LOG_LEVEL || 'info',
-    prettyPrint: process.env.PINO_PRETTY_PRINT === 'true',
-    redact: process.env.PINO_REDACT_PATHS
-        ? redactPathsDefaults.concat(process.env.PINO_REDACT_PATHS.split(','))
+    prettyPrint: process.env.LOG_PRETTY_PRINT === 'true',
+    redact: process.env.LOG_REDACT_PATHS
+        ? redactPathsDefaults.concat(process.env.LOG_REDACT_PATHS.split(','))
         : redactPathsDefaults,
     serializers: pino.stdSerializers
 };