+++ packages/log/CHANGELOG.md
@@ -5,6 +5,8 @@
 -   Dependency upgrades.
 -   Deprecated `insightops` from this package (see @idearium/log-insightops).
 -   Deprecated `LOG_REMOTE`.
+-   Renamed `PINO_PRETTY_PRINT` to `LOG_PRETTY_PRINT`.
+-   Renamed `PINO_REDACT_PATHS` to `LOG_REDACT_PATHS`.
 -   Pretty printing now needs to be explicitly enabled.
 
 ## v1.0.0-beta.1